cloud-security-mcp-server/requirements.txt
# Cloud Security MCP Server Dependencies

# Core MCP Framework
mcp>=1.0.0

# Cloud Provider SDKs
boto3>=1.34.0
botocore>=1.34.0
azure-identity>=1.15.0
azure-mgmt-security>=6.0.0
azure-mgmt-compute>=30.0.0
azure-mgmt-storage>=21.0.0
azure-mgmt-network>=25.0.0
google-cloud-securitycenter>=1.26.0
google-cloud-asset>=3.24.0
google-cloud-compute>=1.15.0
google-cloud-storage>=2.10.0

# Kubernetes
kubernetes>=28.1.0
pyyaml>=6.0.1

# Container & Docker
docker>=7.0.0
python-docker>=1.0.0

# Async & HTTP
aiohttp>=3.9.0
aiofiles>=23.2.0
asyncio-throttle>=1.0.2

# Data Processing
pandas>=2.1.0
numpy>=1.24.0
jinja2>=3.1.0

# Security Tools Integration
subprocess32>=3.5.4
paramiko>=3.4.0
fabric>=3.2.0

# Utilities
pydantic>=2.5.0
rich>=13.7.0
click>=8.1.0
python-dotenv>=1.0.0
cryptography>=41.0.0
requests>=2.31.0
urllib3>=2.0.0

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0
black>=23.0.0
flake8>=6.1.0
mypy>=1.7.0

# Optional: Enhanced features
sqlalchemy>=2.0.0
redis>=5.0.0
celery>=5.3.0
prometheus-client>=0.19.0
